🚚 move SQLAlchemyRule to inner module
