+++ py/src/validb-use/example.py
@@ -12,7 +12,7 @@ from validb import (
     validate_db,
 )
 from validb.datasources import SQLAlchemyDataSource
-from validb.rules import SQLAlchemyRule
+from validb.rules.sqlalchemy import SQLAlchemyRule
 from validb.csvmapping import SimpleDetectionCsvMapping
