+++ py/src/validb/rules/__init__.py
@@ -1,9 +1,6 @@
 from ._rule import Rule, DEFAULT_LEVEL
-from ._sqlalchemy import SQLAlchemyRule, SimpleSQLAlchemyRule
 
 __all__ = [
     "DEFAULT_LEVEL",
     "Rule",
-    "SQLAlchemyRule",
-    "SimpleSQLAlchemyRule",
 ]
