+++ py/src/validb/rules/sqlalchemy/_rule.py
@@ -2,12 +2,12 @@ import typing as t
 
 from sqlalchemy.sql import text
 
-from ..datasources import DataSources, SQLAlchemyDataSource
-from .._embedder import Embedder
-from .._embedded_vars import EmbeddedVariables
-from .._detected import ID, MSG, DETECTION_TYPE, Detected, DetectedType
-from ._rule import Rule, DEFAULT_LEVEL
-from ..formatter import MessageFormatter
+from ...datasources import DataSources, SQLAlchemyDataSource
+from ..._embedder import Embedder
+from ..._embedded_vars import EmbeddedVariables
+from ..._detected import ID, MSG, DETECTION_TYPE, Detected, DetectedType
+from .._rule import Rule, DEFAULT_LEVEL
+from ...formatter import MessageFormatter
 
 
 class SQLAlchemyRule(t.Generic[ID, DETECTION_TYPE, MSG], Rule[ID, DETECTION_TYPE, MSG]):
